feat(ui): add CTRL+N and ctrl+P keybinds
Using (neo)vim in combination with tmux conflicts with the keybindings
ctrl+j and ctrl+k.
The aforementioned bindings are used to move between panes.
ctrl+n and ctrl+p is used in (neo)vim to move to the next end previous
match, which would also make sense in this case.